Splits tests into Unit and Integration
As per Aditi's suggestion in email. Sloppy work right now, will clean
up.
